requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
cssselect2==0.2.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
icalendar==4.0.2
idna==2.7
lxml==4.3.2
oauthlib==2.1.0
Pillow==5.4.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
reportlab==3.5.13
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.11.0
svglib==0.9.0
tinycss2==1.0.0
uritemplate==3.0.0
urllib3==1.23
webencodings==0.5.1
